Getting started with better implementation of activation code checking
